Removes the old business logic ApplicationTree and affilated files, replaces with current data structures.

+++ src/Umbraco.Core/Services/ApplicationTreeService.cs
@@ -158,7 +158,7 @@ namespace Umbraco.Core.Services
         /// <param name="iconClosed">The icon closed.</param>
         /// <param name="iconOpened">The icon opened.</param>
         /// <param name="type">The type.</param>
-        public void MakeNew(bool initialize, byte sortOrder, string applicationAlias, string alias, string title, string iconClosed, string iconOpened, string type)
+        public void MakeNew(bool initialize, int sortOrder, string applicationAlias, string alias, string title, string iconClosed, string iconOpened, string type)
         {
             LoadXml(doc =>
             {
